requirements.txt
attrs==19.3.0
certifi==2019.11.28
Click==7.0
cycler==0.10.0
Flask==1.1.1
Flask-Cors==3.0.8
importlib-metadata==1.4.0
Flask-MonitoringDashboard==3.0.6
itsdangerous==1.1.0
Jinja2==2.11.0
joblib==0.14.1
jsonschema==3.2.0
kiwisolver==1.1.0
kneed==0.5.1
MarkupSafe==1.1.1
matplotlib==3.1.2
more-itertools==8.1.0
mysql-connector-python==8.0.28
numpy==1.18.1
pandas==0.25.3
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
regexp==0.1
scikit-learn==0.22.1
six==1.14.0
sklearn==0.0
Werkzeug==0.16.1
wincertstore==0.2
xgboost==0.90
zipp==2.0.1
